# Remove closure allocations in two methods